feat: add delete endpoint
