+++ src/app/api/journal/route.ts
@@ -15,7 +15,6 @@ export async function GET(req: Request) {
     const session = await getServerSession(authOptions);
 
     if (!session) {
-      console.log("error: Session not found");
       return new NextResponse(
         JSON.stringify({ message: "error: Session not found" }),
         {
@@ -31,7 +30,6 @@ export async function GET(req: Request) {
     const userId = session.user?.id;
 
     if (!userId) {
-      console.log("error: user id not found");
       return new NextResponse(
         JSON.stringify({ message: "error: user id not found" }),
         {
@@ -86,7 +84,6 @@ export async function POST(req: Request, res: Response) {
     const session = await getServerSession(authOptions);
 
     if (!session) {
-      console.log("error: Session not found");
       return new NextResponse(
         JSON.stringify({ message: "error: Session not found" }),
         {
@@ -101,7 +98,6 @@ export async function POST(req: Request, res: Response) {
     const userId = session.user?.id;
 
     if (!userId) {
-      console.log("error: user id not found");
       return new NextResponse(
         JSON.stringify({ message: "error: user id not found" }),
         {
@@ -114,7 +110,6 @@ export async function POST(req: Request, res: Response) {
     }
 
     const json = await req.json();
-    console.log("inner", json);
 
     const { title, contents } = json as CreateJournalEntryRequest;
 
@@ -160,9 +155,104 @@ export async function POST(req: Request, res: Response) {
       }
     );
   } catch (error) {
-    console.error("Internal Server Error:", error);
     return new NextResponse(
       JSON.stringify({ message: "error: Internal server error" + error }),
+      {
+        status: 500,
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }
+    );
+  }
+}
+
+/**
+ * Deletes a journal entry by id, after checking if that id belongs to the user
+ * 400: journal entry id is required
+ * 401: unauthorized or no user found
+ * 404: journal entry not found or does not belong to the user
+ * 500: Internal server error
+ * 200: { message: "success: journal entry deleted", data: JournalEntry }
+ */
+export async function DELETE(req: Request) {
+  try {
+    const url = new URL(req.url);
+    const journalEntryId = url.searchParams.get("id");
+
+    if (!journalEntryId) {
+      return new NextResponse(
+        JSON.stringify({ message: "error: journal entry id is required" }),
+        {
+          status: 400,
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+    }
+
+    const session = await getServerSession(authOptions);
+
+    if (!session || !session.user || !session.user.id) {
+      return new NextResponse(
+        JSON.stringify({ message: "error: unauthorized" }),
+        {
+          status: 401,
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+    }
+
+    const userId = session.user.id;
+
+    // Check if the JournalEntry exists and is associated with the user
+    const journalEntry = await prisma.journalEntry.findFirst({
+      where: {
+        id: journalEntryId,
+        userId: userId, // Ensure the JournalEntry belongs to the specified user
+      },
+    });
+
+    if (!journalEntry) {
+      return new NextResponse(
+        JSON.stringify({
+          message:
+            "error: journal entry not found or does not belong to the user",
+        }),
+        {
+          status: 404,
+          headers: {
+            "Content-Type": "application/json",
+          },
+        }
+      );
+    }
+
+    // Proceed with deletion if the JournalEntry exists and is associated with the user
+    const deletedJournalEntry = await prisma.journalEntry.delete({
+      where: {
+        id: journalEntryId,
+      },
+    });
+
+    return new NextResponse(
+      JSON.stringify({
+        message: "success: journal entry deleted",
+        data: deletedJournalEntry,
+      }),
+      {
+        status: 200,
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }
+    );
+  } catch (error) {
+    return new NextResponse(
+      JSON.stringify({ message: "error: failed to delete journal entry" }),
       {
         status: 500,
         headers: {
