+++ src/app/journals/page.tsx
@@ -2,24 +2,21 @@
 import { JournalEntry, Sender, TextObject } from "@/types";
 import { useSession } from "next-auth/react";
 import React, { useEffect, useState } from "react";
+import { Trash2 } from "lucide-react";
 
 type Props = {};
 
 const page = (props: Props) => {
   const { status } = useSession();
   const [journals, setJournals] = useState<JournalEntry[] | null>(null);
-  useEffect(() => {
+
+  const fetchJournals = () => {
     fetch("/api/journal")
       .then((res) => res.json())
       .then((data) => {
-        console.log(data);
         setJournals(data.data);
       });
-  }, []);
-
-  if (status === "loading") {
-    return <div>Loading...</div>;
-  }
+  };
 
   const handleSubmit = async () => {
     const contents: TextObject[] = [
@@ -58,21 +55,54 @@ const page = (props: Props) => {
       },
       body: JSON.stringify({ title: "My journal", contents: contents }),
     });
-
-    console.log("response", response);
+    fetchJournals();
   };
 
+  const handleDelete = async (id: string) => {
+    const response = await fetch(
+      `/api/journal?` +
+        new URLSearchParams({
+          id: id,
+        }).toString(),
+      {
+        method: "DELETE",
+      }
+    );
+
+    fetchJournals();
+  };
+
+  useEffect(() => {
+    fetchJournals();
+  }, []);
+
+  if (status === "loading") {
+    return <div>Loading...</div>;
+  }
+
   return (
-    <div className="flex flex-col min-h-screen box-border p-8 space-y-4">
+    <div className="flex flex-col min-h-screen w-full box-border p-8 space-y-4">
       <h1 className="font-semibold text-lg">Journals</h1>
-      <div className="border border-gray-300 rounded-sm p-4 box-border w-full">
+      <div className="flex flex-col space-y-4 w-full">
         {journals &&
           journals.map((journal) => (
-            <div key={journal.id}>
-              <h1>Title: {journal.title}</h1>
-              <span>
-                Content: {journal.contents[0] && journal.contents[0].content}
-              </span>
+            <div
+              key={journal.id}
+              className="flex space-x-4 border border-gray-300 rounded-sm p-4 box-border"
+            >
+              <div className="flex flex-grow flex-col">
+                <h1 className="">{journal.title}</h1>
+                <span>
+                  Content: {journal.contents[0] && journal.contents[0].content}
+                </span>
+              </div>
+              <button
+                onClick={() => {
+                  handleDelete(journal.id);
+                }}
+              >
+                <Trash2 />
+              </button>
             </div>
           ))}
       </div>
